Put gameboard into seperate file for easier reusablility.
