+++ CMakeLists.txt
@@ -25,7 +25,7 @@ FetchContent_MakeAvailable(raylib)
 
 
 # Add source to this project's executable.
-add_executable (Minesweeper "Minesweeper.cpp" )
+add_executable (Minesweeper "Minesweeper.cpp"  "GameBoard.h")
 
 if (CMAKE_VERSION VERSION_GREATER 3.12)
   set_property(TARGET Minesweeper PROPERTY CXX_STANDARD 20)
